WIP refactor subcase screen to facilitate other subcase types
