+++ app/components/subcase/detail/bekrachtiging.hbs
@@ -1,0 +1,1 @@
+TODO MAKE THIS SCREEN
