+++ app/components/subcase/detail/panels/government-areas.hbs
@@ -1,0 +1,11 @@
+<Utils::GovernmentAreas::GovernmentAreasPanel
+  @referenceDate={{or
+    this.meeting.plannedStart
+    this.agenda.created
+    @subcase.created
+  }}
+  @governmentAreas={{@subcase.governmentAreas}}
+  @allowEditing={{user-may "manage-cases"}}
+  @onSave={{this.saveGovernmentAreas}}
+  class="auk-u-mb-8"
+/>
