+++ app/components/subcase/detail/panels/mandatees.hbs
@@ -1,0 +1,11 @@
+<Mandatees::MandateesPanel
+  @referenceDate={{if @meeting @meeting.plannedStart @subcase.created}}
+  @mandatees={{@mandatees}}
+  @submitter={{@submitter}}
+  @emptyStateMessage={{if
+    this.showMandateesNotApplicableMessage
+    (t "not-applicable")
+  }}
+  @allowEditing={{user-may "manage-cases"}}
+  @onSave={{this.saveMandateeData}}
+/>
