+++ translations/nl-be.json
@@ -1338,5 +1338,6 @@
   "updated-on": "Upgedate op",
   "on-this-page": "Op deze pagina",
   "design-decreet-of": "Ontwerp van decreet over {subcaseShortTitle}",
-  "search-not-allowed-during-reordering": "Zoeken is niet mogelijk tijdens het aanpassen van de volgorde"
+  "search-not-allowed-during-reordering": "Zoeken is niet mogelijk tijdens het aanpassen van de volgorde",
+  "ratification-overview": "Bekrachtiging overzicht"
 }
